인벤토리 구현 끝
더블클릭 사용 구현

+++ RSP/Source/RSP/Private/RSP_StatComponent.cpp
@@ -43,7 +43,7 @@ void URSP_StatComponent::TickComponent(float DeltaTime, ELevelTick TickType, FAc
 	// ...
 }
 
-void URSP_StatComponent::AddCurHp(float amount)
+void URSP_StatComponent::AddCurHp(int32 amount)
 {
 	int32 before = _curHp;
